% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_drafts.R
\name{tidy_drafts}
\alias{tidy_drafts}
\title{Get a tidy dataset of NHL entry drafts}
\usage{
tidy_drafts(
  drafts_year = NULL,
  keep_id = FALSE,
  return_datatable = getOption("tidynhl.data.table", TRUE)
)
}
\arguments{
\item{drafts_year}{\emph{(optional)}  Integer vector indicating which entry drafts will be returned.
The first NHL entry draft was held in 1963. Default to \code{NULL} which will return every entry
drafts in NHL history.}

\item{keep_id}{\emph{(optional)} Logical indicating if the IDs of different dimensions should be
returned. Default to \code{FALSE}.}

\item{return_datatable}{\emph{(optional)} Logical indicating whether or not a data.table should be
returned. Default can be set globally with \code{options("tidynhl.data.table")}.}
}
\description{
The function \code{tidy_drafts()} is meant to be a user-friendly way of getting data about the NHL
entry drafts.
}
\examples{
# Allowing large outputs for the pkgdown website
options(width = 1000L)

# Get every drafts in NHL history
tidy_drafts()

# Get both the 2019 and 2020 NHL entry drafts, keeping  the IDs
tidy_drafts(drafts_year = 2019:2020, keep_id = TRUE)

}
