% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_games_events.R
\name{tidy_games_events}
\alias{tidy_games_events}
\title{Get a tidy dataset of NHL games events}
\usage{
tidy_games_events(
  games_id,
  time_elapsed = TRUE,
  standardized_coordinates = TRUE,
  keep_id = FALSE,
  return_datatable = getOption("tidynhl.data.table", TRUE)
)
}
\arguments{
\item{games_id}{Integer vector indicating the NHL ID of selected games for which the events will
be returned. The required format is 'xxxxyyzzzz' where
\itemize{
\item 'xxxx' is the first 4 digits of a valid NHL season ID,
\item 'yy' is one of "02" (regular) or "03" (playoffs),
\item 'zzzz' is a 4-digit number attributed to a single game.
}}

\item{time_elapsed}{\emph{(optional)} Logical indicating if the time sould be indicated as elapsed
(\code{TRUE}) or remaining (\code{FALSE}). Default to \code{TRUE}.}

\item{standardized_coordinates}{\emph{(optional)} Logical indicating if the ice coordinates should be
standardized (away team zone negative on the x-axis) or kept as is. Default to \code{TRUE}.}

\item{keep_id}{\emph{(optional)} Logical indicating if the IDs of different dimensions should be
returned. However, note that \code{game_id} will always be returned. Default to \code{FALSE}.}

\item{return_datatable}{\emph{(optional)} Logical indicating whether or not a data.table should be
returned. Default can be set globally with \code{options("tidynhl.data.table")}.}
}
\description{
The function \code{tidy_games_events()} is meant to be a user-friendly way of connecting events of
different types in selected NHL games.
}
\examples{
# Allowing large outputs for the pkgdown website
options(width = 1000L)

# Get events of the 2021-01-13 MTL @ TOR game
tidy_games_events(2020020003L)

# Get events of both the 2021-01-13 MTL @ TOR and 2021-01-14 BOS @ NJD games with time remaining
# instead of time elapsed, as is coordinates instead of normalized, and keeping the IDs
tidy_games_events(
  games_id = c(2020020003L, 2020020007L),
  time_elapsed = FALSE,
  standardized_coordinates = FALSE,
  keep_id = TRUE
)

}
