% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_players_meta.R
\name{tidy_players_meta}
\alias{tidy_players_meta}
\title{Get a tidy dataset of NHL players metadata}
\usage{
tidy_players_meta(
  players_id = NULL,
  keep_id = FALSE,
  return_datatable = getOption("tidynhl.data.table", TRUE)
)
}
\arguments{
\item{players_id}{\emph{(optional)} Integer vector indicating the NHL ID of players for whom the
metadata will be returned. Default to \code{NULL} which will return every players in the NHL
database.}

\item{keep_id}{\emph{(optional)} Logical indicating if the IDs of different dimensions should be
returned. Default to \code{FALSE}.}

\item{return_datatable}{\emph{(optional)} Logical indicating whether or not a data.table should be
returned. Default can be set globally with \code{options("tidynhl.data.table")}.}
}
\description{
The function \code{tidy_players_meta()} is meant to be a user-friendly way of getting metadata about
the selected players.
}
\examples{
# Allowing large outputs for the pkgdown website
options(width = 1000L)

# Get metadata of every players in the NHL database
tidy_players_meta()

# Get metadata about both Vincent Lecavalier and Carey Price, keeping the IDs
tidy_players_meta(players_id = c(8467329L, 8471679L), keep_id = TRUE)

}
