% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_players_stats.R
\name{tidy_players_stats}
\alias{tidy_players_stats}
\title{Get a tidy dataset of NHL players statistics}
\usage{
tidy_players_stats(
  players_id,
  regular = TRUE,
  playoffs = TRUE,
  keep_id = FALSE,
  return_datatable = NULL
)
}
\arguments{
\item{players_id}{Integer vector indicating the NHL ID of players for whom the statistics will be returned.}

\item{regular}{\emph{(optional)} Logical indicating if the regular season statistics should be returned. Default to \code{TRUE}.}

\item{playoffs}{\emph{(optional)} Logical indicating if the playoffs statistics should be returned. Default to \code{TRUE}.}

\item{keep_id}{\emph{(optional)} Logical indicating if the IDs of different dimensions should be returned. Default to \code{FALSE}.}

\item{return_datatable}{\emph{(optional)} Logical indicating whether or not a data.table should be returned. Default to \code{TRUE} if the \code{data.table} package is
attached in the active session.}
}
\description{
The function \code{tidy_players_stats()} is meant to be a user-friendly way of getting the NHL year-by-year statistics of selected players.
}
\examples{
# Load the data.table package to easily manipulate the data
library(data.table)

# Get the NHL career statistics of Vincent Lecavalier
stats_lecavalier <- tidy_players_stats(8467329L)

# Print the column names
colnames(stats_lecavalier)

# Print an excerpt of the data
stats_lecavalier[, .(season_years, season_type, team_abbreviation, skater_games, skater_points)]

# Get the NHL career playoffs statistics of both Vincent Lecavalier and Carey Price, keeping the IDs
stats_playoffs_lecavalier_price <- tidy_players_stats(
  players_id = c(8467329L, 8471679L),
  regular    = FALSE,
  keep_id    = TRUE
)

# Print the column names
colnames(stats_playoffs_lecavalier_price)

# Print an excerpt of the data
stats_playoffs_lecavalier_price[, .(player_id, player_name, season_years, skater_points, goalie_savepct)]

}
