% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_schedule.R
\name{tidy_schedule}
\alias{tidy_schedule}
\title{Get a tidy dataset of the NHL schedule}
\usage{
tidy_schedule(
  seasons_id,
  regular = TRUE,
  playoffs = TRUE,
  tz = Sys.timezone(),
  keep_id = FALSE,
  return_datatable = getOption("tidynhl.data.table", TRUE)
)
}
\arguments{
\item{seasons_id}{Character vector of the seasons ID for which the schedule will be returned. The
required format is 'xxxxyyyy'.}

\item{regular}{\emph{(optional)} Logical indicating if the regular season schedule should be returned.
Default to \code{TRUE}.}

\item{playoffs}{\emph{(optional)} Logical indicating if the playoffs schedule should be returned.
Default to \code{TRUE}.}

\item{tz}{\emph{(optional)} Character specifying the timezone that should be used for datetime.
Default to the user system timezone.}

\item{keep_id}{\emph{(optional)} Logical indicating if the IDs of different dimensions should be
returned. Default to \code{FALSE}.}

\item{return_datatable}{\emph{(optional)} Logical indicating whether or not a data.table should be
returned. Default can be set globally with options("tidynhl.data.table").}
}
\description{
The function \code{tidy_schedule()} is meant to be a user-friendly way of getting the NHL schedule,
including the final score of completed games.
}
\examples{
# Allowing large outputs for the pkgdown website
options(width = 1000L)

# Get the schedule of the 2019-2020 regular season and playoffs
tidy_schedule("20192020")

# Get the regular season schedule of both the 2018-2019 and 2019-2020 seasons, keeping the IDs
# and indicating game datetime with Los Angeles local time
tidy_schedule(
  seasons_id = c("20182019", "20192020"),
  playoffs = FALSE,
  tz = "America/Los_Angeles",
  keep_id = TRUE
)

}
