% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_seasons_meta.R
\name{tidy_seasons_meta}
\alias{tidy_seasons_meta}
\title{Get a tidy dataset of NHL seasons metadata}
\usage{
tidy_seasons_meta(
  keep_id = FALSE,
  return_datatable = getOption("tidynhl.data.table", TRUE)
)
}
\arguments{
\item{keep_id}{\emph{(optional)} Logical indicating if the IDs of different dimensions should be
returned. Default to \code{FALSE}.}

\item{return_datatable}{\emph{(optional)} Logical indicating whether or not a data.table should be
returned. Default can be set globally with \code{options("tidynhl.data.table")}.}
}
\description{
The function \code{tidy_seasons_meta()} is meant to be a user-friendly way of getting metadata about
NHL seasons.
}
\examples{
# Allowing large outputs for the pkgdown website
options(width = 1000L)

# Get metadata of NHL seasons
tidy_seasons_meta()

}
