% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_skaters_gamelogs.R
\name{tidy_skaters_gamelogs}
\alias{tidy_skaters_gamelogs}
\title{Get a tidy dataset of NHL skaters game logs}
\usage{
tidy_skaters_gamelogs(
  players_id,
  seasons_id = NULL,
  regular = TRUE,
  playoffs = TRUE,
  tz = Sys.timezone(),
  keep_id = FALSE,
  return_datatable = getOption("tidynhl.data.table", TRUE)
)
}
\arguments{
\item{players_id}{Integer vector indicating the NHL ID of skaters for whom the game logs will be
returned.}

\item{seasons_id}{\emph{(optional)} Character vector indicating the seasons for which the game logs
will be returned. The required format for each character is 'xxxxyyyy'. Default to \code{NULL} which
will return the full career game logs for each player.}

\item{regular}{\emph{(optional)} Logical indicating if the regular season game logs should be
returned. Default to \code{TRUE}.}

\item{playoffs}{\emph{(optional)} Logical indicating if the playoffs game logs should be returned.
Default to \code{TRUE}.}

\item{tz}{\emph{(optional)} Character specifying the timezone that should be used for datetime.
Default to the user system timezone.}

\item{keep_id}{\emph{(optional)} Logical indicating if the IDs of different dimensions should be
returned. Default to \code{FALSE}.}

\item{return_datatable}{\emph{(optional)} Logical indicating whether or not a data.table should be
returned. Default can be set globally with \code{options("tidynhl.data.table")}.}
}
\description{
The function \code{tidy_skaters_gamelogs()} is meant to be a user-friendly way of getting the NHL game
logs of selected skaters.
}
\examples{
# Allowing large outputs for the pkgdown website
options(width = 1000L)

# Get the NHL career game logs of Vincent Lecavalier
tidy_skaters_gamelogs(8467329L)

# Get the 2010-11 NHL playoffs game logs of both Vincent Lecavalier and Martin St. Louis,
# keeping the IDs
tidy_skaters_gamelogs(
  players_id = c(8467329L, 8466378L),
  seasons_id = "20102011",
  regular = FALSE,
  keep_id = TRUE
)

}
