% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_skaters_stats.R
\name{tidy_skaters_stats}
\alias{tidy_skaters_stats}
\title{Get a tidy dataset of NHL skaters statistics}
\usage{
tidy_skaters_stats(
  players_id,
  regular = TRUE,
  playoffs = TRUE,
  keep_id = FALSE,
  return_datatable = getOption("tidynhl.data.table", TRUE)
)
}
\arguments{
\item{players_id}{Integer vector indicating the NHL ID of skaters for whom the statistics will be
returned.}

\item{regular}{\emph{(optional)} Logical indicating if the regular season statistics should be
returned. Default to \code{TRUE}.}

\item{playoffs}{\emph{(optional)} Logical indicating if the playoffs statistics should be returned.
Default to \code{TRUE}.}

\item{keep_id}{\emph{(optional)} Logical indicating if the IDs of different dimensions should be
returned. Default to \code{FALSE}.}

\item{return_datatable}{\emph{(optional)} Logical indicating whether or not a data.table should be
returned. Default can be set globally with \code{options("tidynhl.data.table")}.}
}
\description{
The function \code{tidy_skaters_stats()} is meant to be a user-friendly way of getting the NHL
year-by-year statistics of selected skaters.
}
\examples{
# Allowing large outputs for the pkgdown website
options(width = 1000L)

# Get the NHL career statistics of Vincent Lecavalier
tidy_skaters_stats(8467329L)

# Get the NHL career playoffs statistics of both Vincent Lecavalier and Martin St. Louis, keeping
# the IDs
tidy_skaters_stats(c(8467329L, 8466378L), regular = FALSE, keep_id = TRUE)

}
