% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_teams_meta.R
\name{tidy_teams_meta}
\alias{tidy_teams_meta}
\title{Get a tidy dataset of NHL teams metadata}
\usage{
tidy_teams_meta(
  active_only = TRUE,
  keep_id = FALSE,
  return_datatable = getOption("tidynhl.data.table", TRUE)
)
}
\arguments{
\item{active_only}{\emph{(optional)} Logical indicating if only active teams should be returned.
Default to \code{TRUE}.}

\item{keep_id}{\emph{(optional)} Logical indicating if the IDs of different dimensions should be
returned. Default to \code{FALSE}.}

\item{return_datatable}{\emph{(optional)} Logical indicating whether or not a data.table should be
returned. Default can be set globally with \code{options("tidynhl.data.table")}.}
}
\description{
The function \code{tidy_teams_meta()} is meant to be a user-friendly way of getting metadata about NHL
teams.
}
\examples{
# Allowing large outputs for the pkgdown website
options(width = 1000L)

# Get metadata of active teams
tidy_teams_meta()

# Get metadata of active and historical teams keeping the IDs
tidy_teams_meta(active_only = FALSE, keep_id = TRUE)

}
