% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_draft.R
\name{tidy_draft}
\alias{tidy_draft}
\title{Get a tidy dataset of the NHL entry drafts}
\usage{
tidy_draft(
  drafts_year,
  keep_id = FALSE,
  return_datatable = getOption("tidynhl.data.table", TRUE)
)
}
\arguments{
\item{drafts_year}{Integer vector indicating which entry drafts will be returned. The first NHL
entry draft was held in 1963.}

\item{keep_id}{\emph{(optional)} Logical indicating if the IDs of different dimensions should be
returned. Default to \code{FALSE}.}

\item{return_datatable}{\emph{(optional)} Logical indicating whether or not a data.table should be
returned. Default can be set globally with options("tidynhl.data.table").}
}
\description{
The function \code{tidy_draft()} is meant to be a user-friendly way of getting data about the NHL
entry drafts.
}
\examples{
# Allowing large outputs for the pkgdown website
options(width = 1000L)

# Get the 2020 NHL entry draft
tidy_draft(2020L)

# Get both the 2019 and 2020 NHL entry drafts, keeping  the IDs
tidy_draft(drafts_year = 2019:2020, keep_id = TRUE)

}
