% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_draft.R
\name{tidy_draft}
\alias{tidy_draft}
\title{Get a tidy dataset of the NHL entry drafts}
\usage{
tidy_draft(drafts_year, keep_id = FALSE, return_datatable = NULL)
}
\arguments{
\item{drafts_year}{Integer vector indicating which entry drafts will be returned. The first NHL entry draft was held in 1963.}

\item{keep_id}{\emph{(optional)} Logical indicating if the IDs of different dimensions should be returned. Default to \code{FALSE}.}

\item{return_datatable}{\emph{(optional)} Logical indicating whether or not a data.table should be returned. Default to \code{TRUE} if the \code{data.table} package is
attached in the active session.}
}
\description{
The function \code{tidy_draft()} is meant to be a user-friendly way of getting data about the NHL entry drafts.
}
\examples{
# Get the 2019 NHL entry draft
head(tidy_draft(2019L))

# Get both the 2018 and 2019 NHL entry drafts, keeping  the ID
head(tidy_draft(
  drafts_year = 2018:2019,
  keep_id     = TRUE
))

}
