% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_players_gamelogs.R
\name{tidy_players_gamelogs}
\alias{tidy_players_gamelogs}
\title{Get a tidy dataset of NHL players game logs}
\usage{
tidy_players_gamelogs(
  players_id,
  seasons_id = NULL,
  regular = TRUE,
  playoffs = TRUE,
  tz = Sys.timezone(),
  keep_id = FALSE,
  return_datatable = NULL
)
}
\arguments{
\item{players_id}{Integer vector indicating the NHL ID of players for whom the game logs will be returned.}

\item{seasons_id}{\emph{(optional)} Character vector indicating the season for which the game logs will be returned. The required format for each character is
'xxxxyyyy'. Default to \code{NULL} which will return the full career game logs for each player.}

\item{regular}{\emph{(optional)} Logical indicating if the regular season schedule should be returned. Default to \code{TRUE}.}

\item{playoffs}{\emph{(optional)} Logical indicating if the playoffs schedule should be returned. Default to \code{TRUE}.}

\item{tz}{\emph{(optional)} Character specifying the timezone that should be used for datetime. Default to the user system timezone.}

\item{keep_id}{\emph{(optional)} Logical indicating if the IDs of different dimensions should be returned. Default to \code{FALSE}.}

\item{return_datatable}{\emph{(optional)} Logical indicating whether or not a data.table should be returned. Default to \code{TRUE} if the \code{data.table} package is
attached in the active session.}
}
\description{
The function \code{tidy_players_gamelogs()} is meant to be a user-friendly way of getting the NHL game logs of selected players.
}
\examples{
# Get the NHL career game logs of Vincent Lecavalier
head(tidy_players_gamelogs(8467329L))

# Get the NHL 2010-11 playoffs statistics of both Vincent Lecavalier and Carey Price, keeping the IDs
tidy_players_gamelogs(
  players_id = c(8467329L, 8471679L),
  seasons_id = "20102011",
  regular    = FALSE,
  keep_id    = TRUE
)

}
