% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidynhl-package.R
\docType{package}
\name{tidynhl-package}
\alias{tidynhl}
\alias{tidynhl-package}
\title{tidynhl: An Easy Way to Extract NHL Data in a Tidy Format}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

A user-friendly package which aims at facilitating data extraction through the NHL API in a ready-to-use format.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://jplecavalier.github.io/tidynhl}
  \item \url{https://github.com/jplecavalier/tidynhl}
}

}
\author{
\strong{Maintainer}: J.P. Le Cavalier \email{jplecavalier@me.com} (\href{https://orcid.org/0000-0001-5904-2641}{ORCID})

}
\keyword{internal}
